% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufos.R
\name{.onLoad}
\alias{.onLoad}
\title{Finalizer for the entire UFO framework, called on session exit.}
\usage{
.onLoad(libname, pkgname)
}
\description{
Finalizer for the entire UFO framework, called on session exit.
}
