% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufos.R
\name{.onUnload}
\alias{.onUnload}
\title{Finalizer for the entire UFO framework, called on package unload.
AFAIK this cannot be relied on to be called on shutdown.}
\usage{
.onUnload(libname, pkgname)
}
\description{
Finalizer for the entire UFO framework, called on package unload.
AFAIK this cannot be relied on to be called on shutdown.
}
