% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufos.R
\docType{data}
\name{.ufo_core}
\alias{.ufo_core}
\title{A UFO core singleton that is lazily loaded and hangs around until the
package is unloaded.}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
.ufo_core
}
\description{
A UFO core singleton that is lazily loaded and hangs around until the
package is unloaded.
}
\keyword{datasets}
