% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{ufo_integer}
\alias{ufo_integer}
\title{Creates a UFO object of type \code{integer}}
\usage{
ufo_integer(length = 0, initial_value = as.integer(0), ...)
}
\arguments{
\item{length}{the length fo the ufo integer vector, has to be larger than 1}

\item{initial_value}{the intial value of the elements in the vector
(0 by default)}

\item{...}{other ufo configuration options (see \code{ufo_integer_constructor})}

\item{read_only}{sets the vector to be write-protected by the OS
(optional, false by default).}

\item{chunk_length}{the minimum number of elements loaded at once,
will always be rounded up to a full memory page
(optional, a page by default).}
}
\value{
a ufo vector of size \code{length}, lazily populated with \code{initial_value}
}
\description{
Creates a UFO object of type \code{integer}
}
