% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufos.R
\name{ufo_system_get_or_create}
\alias{ufo_system_get_or_create}
\title{Get a reference to the UFO core.}
\usage{
ufo_system_get_or_create(writeback_path, high_water_mark, low_water_mark)
}
\arguments{
\item{writeback_path}{location of directory where UFOs write modified
dematerialized data}

\item{high_water_mark}{maximum size of materialized UFO chunks (in MB)}

\item{low_water_mark}{size of materialized UFO chunks after GC (in MB)}
}
\value{
An external pointer object with functions attached.
}
\description{
Get a reference to the UFO core.
}
